requirements.txt
Django
django-cors-headers==3.8.0
djangorestframework==3.12.4
